Completed Google Authorization/Authentication UI and Functionality

+++ controllers/api/v1/posts_api.js
@@ -23,13 +23,19 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id){
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
-    return res.json(200, {
-      message: 'post and all asociated comments are deleted successfully'
-    });
+      await Comment.deleteMany({ post: req.params.id });
+      return res.json(200, {
+        message: 'post and all asociated comments are deleted successfully'
+      });
+    }
+    else {
+      return res.json(401, {
+        message: "You cannot delete this post"
+      })
+    }
   }
   catch (err) {
     console.log(err)
